+++ lib/ui/groups/GroupPostDetailPanel.dart
@@ -64,7 +64,7 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
   //Main Post - Edit/Show
   Post? _post; //Main post {Data Presentation}
   List<Comment>? _replies; //Main post comments
-  PostDataModel? _mainPostUpdateData;//Main Post Edit
+  PostUpdateData? _mainPostUpdateData;//Main Post Edit
   List<Member>? _allMembersAllowedToPost;
   //Reply - Edit/Create/Show
   Comment? _focusedReply; //Focused on Reply {Replies Thread Presentation} // User when Refresh post thread
@@ -283,6 +283,26 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
                                                     width: 0.0))),
                                         style: Styles().textStyles.getTextStyle("widget.input_field.text.regular"),
                                        )),
+                                Visibility(visible: _isPost && _canPinPost,
+                                    child: Container(
+                                      padding: EdgeInsets.only(top: 8, bottom: _outerPadding),
+                                      child: EnabledToggleButton(
+                                          label: "Pin post to top of all posts (Only one pinned post per group is allowed. Pinning this post will automatically unpin any past admin posts.)",
+                                          borderRadius: BorderRadius.circular(4),
+                                          border: Border.all(color: Styles().colors.surfaceAccent, width: 1),
+                                          toggled: _mainPostUpdateData?.pinned == true,
+                                          textStyle: Styles().textStyles.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"),
+                                          enabled: true,
+                                          onTap: () {
+                                            if(mounted){
+                                              setState(() {
+                                                _mainPostUpdateData?.pinned  = !(_mainPostUpdateData?.pinned ?? false);
+                                              });
+                                            }
+                                          }
+                                      ),
+                                    )
+                                ),
                                 Row(children: [
                                   Flexible(
                                       flex: 1,
@@ -701,12 +721,11 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
     if (CollectionUtils.isNotEmpty(selectedAccountIds) && selectedAccountIds!.contains(Auth2().accountId)) {
       selectedAccountIds.remove(Auth2().accountId);
     }
-    _mainPostUpdateData = PostDataModel(
-        body: _post?.body,
-        imageUrl: _post?.imageUrl,
+    _mainPostUpdateData = PostUpdateData.fromPost(_post,
         members: GroupMembersSelectionWidget.constructUpdatedMembersList(
-            selectedAccountIds: selectedAccountIds, upToDateMembers: _allMembersAllowedToPost),
-        dateScheduled: _post?.dateActivatedLocal);
+          selectedAccountIds: selectedAccountIds,
+          upToDateMembers: _allMembersAllowedToPost),
+    );
     setStateIfMounted(() {});
   }
 
@@ -731,8 +750,15 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
     }
     _post!.setMemberAccountIds(groupId: _groupId, accountIds: memberAccountIds);
     Social().updatePost(post: _post!).then((succeeded) {
-      _mainPostUpdateData = null;
-      _setLoading(false);
+      if(_mainPostUpdateData?.pinned != _post?.pinned){
+        Social().pinPost(postId: _post?.id ?? "", pinned: _mainPostUpdateData?.pinned == true).whenComplete((){
+          _mainPostUpdateData = null;
+          _setLoading(false);
+        });
+      } else {
+        _mainPostUpdateData = null;
+        _setLoading(false);
+      }
     });
   }
 
@@ -933,6 +959,12 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
 
   bool get _isReportAbuseVisible => widget.group.currentUserIsMemberOrAdmin;
 
+  bool get _isPost => _post?.type ==  PostType.post;
+
+  bool get _canPinPost => _isAdmin;
+
+  bool get _isAdmin => widget.group.currentUserIsAdmin;
+
   bool get _isEditMainPost => _mainPostUpdateData != null;
 
   String get _groupId => widget.group.id!;
@@ -944,5 +976,21 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
       _refreshPostData();
     }
   }
+}
+
+class PostUpdateData extends PostDataModel {
+  bool? pinned;
+
+  PostUpdateData({String? body, String? subject, String? imageUrl, List<Member>? members, DateTime? dateScheduled, this.pinned}) : 
+        super(body: body, subject: subject, imageUrl: imageUrl, members: members, dateScheduled: dateScheduled);
+  
+  factory PostUpdateData.fromPost(Post? post, {List<Member>? members}) =>
+      PostUpdateData(
+          body: post?.body,
+          imageUrl: post?.imageUrl,
+          dateScheduled: post?.dateActivatedLocal,
+          pinned: post?.pinned,
+          members: members
+      );
 }